+++ pkg/cli/store.go
@@ -29,7 +29,7 @@ var NotFound = errors.New("Flag not found")
 func init() {
 	set := &pflag.FlagSet{}
 
-	// Set of all flags used across GUAC clis and subcommands. Use consistant
+	// Set of all flags used across GUAC clis and subcommands. Use consistent
 	// names for config file.
 	set.String("nats-addr", "nats://127.0.0.1:4222", "address to connect to NATs Server")
 	set.String("csub-addr", "localhost:2782", "address to connect to collect-sub service")
@@ -74,6 +74,9 @@ func init() {
 
 	set.StringP("vuln-id", "v", "", "CVE, GHSA or OSV ID to check")
 	set.Int("num-path", 0, "number of paths to return, 0 means all paths")
+	set.String("start-purl", "", "string input of purl with package to start search from")
+	set.String("stop-purl", "", "string input of purl with package to stop search at")
+	set.Bool("sample-data", false, "ingest sample data")
 
 	// Google Cloud platform flags
 	set.String("gcp-credentials-path", "", "Path to the Google Cloud service account credentials json file.\nAlternatively you can set GOOGLE_APPLICATION_CREDENTIALS=<path> in your environment.")
