+++ pkg/guacanalytics/patchPlanning_test.go
@@ -180,7 +180,7 @@ var (
 					Digest:    "testArtifactDigest4",
 				},
 				IsOccurrence: &model.IsOccurrenceInputSpec{
-					Justification: "connect pkg5 and artifact3",
+					Justification: "connect pkg5 and artifact4",
 				},
 			},
 		},
@@ -930,256 +930,6 @@ var (
 	}
 )
 
-func ingestIsDependency(ctx context.Context, client graphql.Client, graph assembler.IngestPredicates) error {
-	for _, ingest := range graph.IsDependency {
-		_, err := model.IngestPackage(ctx, client, *ingest.Pkg)
-
-		if err != nil {
-			return fmt.Errorf("error in ingesting package: %s\n", err)
-		}
-
-		_, err = model.IngestPackage(ctx, client, *ingest.DepPkg)
-
-		if err != nil {
-			return fmt.Errorf("error in ingesting dependent package: %s\n", err)
-		}
-		_, err = model.IsDependency(ctx, client, *ingest.Pkg, *ingest.DepPkg, *ingest.IsDependency)
-
-		if err != nil {
-			return fmt.Errorf("error in ingesting isDependency: %s\n", err)
-		}
-	}
-	return nil
-}
-
-func ingestHasSLSA(ctx context.Context, client graphql.Client, graph assembler.IngestPredicates) error {
-	for _, ingest := range graph.HasSlsa {
-		_, err := model.IngestBuilder(ctx, client, *ingest.Builder)
-
-		if err != nil {
-			return fmt.Errorf("error in ingesting Builder for HasSlsa: %v\n", err)
-		}
-		_, err = model.SLSAForArtifact(ctx, client, *ingest.Artifact, ingest.Materials, *ingest.Builder, *ingest.HasSlsa)
-
-		if err != nil {
-			return fmt.Errorf("error in ingesting HasSlsa: %v\n", err)
-		}
-	}
-	return nil
-}
-
-func ingestHasSourceAt(ctx context.Context, client graphql.Client, graph assembler.IngestPredicates) error {
-	for _, ingest := range graph.HasSourceAt {
-		_, err := model.IngestPackage(ctx, client, *ingest.Pkg)
-
-		if err != nil {
-			return fmt.Errorf("error in ingesting pkg HasSourceAt: %v\n", err)
-		}
-
-		_, err = model.IngestSource(ctx, client, *ingest.Src)
-
-		if err != nil {
-			return fmt.Errorf("error in ingesting src HasSourceAt: %v\n", err)
-		}
-
-		_, err = model.HasSourceAt(ctx, client, *ingest.Pkg, ingest.PkgMatchFlag, *ingest.Src, *ingest.HasSourceAt)
-
-		if err != nil {
-			return fmt.Errorf("error in ingesting HasSourceAt: %v\n", err)
-		}
-	}
-	return nil
-}
-
-func ingestIsOccurrence(ctx context.Context, client graphql.Client, graph assembler.IngestPredicates) error {
-	for _, ingest := range graph.IsOccurrence {
-		var err error
-
-		if ingest.Src != nil {
-			_, err = model.IngestSource(ctx, client, *ingest.Src)
-		} else {
-			_, err = model.IngestPackage(ctx, client, *ingest.Pkg)
-
-		}
-
-		if err != nil {
-			return fmt.Errorf("error in ingesting pkg/src IsOccurrence: %v\n", err)
-		}
-
-		_, err = model.IngestArtifact(ctx, client, *ingest.Artifact)
-
-		if err != nil {
-			return fmt.Errorf("error in ingesting artifact for IsOccurrence: %v\n", err)
-		}
-
-		if ingest.Src != nil {
-			_, err = model.IsOccurrenceSrc(ctx, client, *ingest.Src, *ingest.Artifact, *ingest.IsOccurrence)
-		} else {
-			_, err = model.IsOccurrencePkg(ctx, client, *ingest.Pkg, *ingest.Artifact, *ingest.IsOccurrence)
-		}
-
-		if err != nil {
-			return fmt.Errorf("error in ingesting isOccurrence: %v\n", err)
-		}
-	}
-	return nil
-}
-
-func ingestCertifyGood(ctx context.Context, client graphql.Client, graph assembler.IngestPredicates) error {
-	for _, ingest := range graph.CertifyGood {
-		_, err := model.IngestPackage(ctx, client, *ingest.Pkg)
-
-		if err != nil {
-			return fmt.Errorf("error in ingesting Package for CertifyGood: %v\n", err)
-		}
-
-		_, err = model.CertifyGoodPkg(ctx, client, *ingest.Pkg, ingest.PkgMatchFlag, *ingest.CertifyGood)
-
-		if err != nil {
-			return fmt.Errorf("error in ingesting CertifyGood: %v\n", err)
-		}
-	}
-	return nil
-}
-
-func ingestPkgEqual(ctx context.Context, client graphql.Client, graph assembler.IngestPredicates) error {
-	for _, ingest := range graph.PkgEqual {
-		_, err := model.IngestPackage(ctx, client, *ingest.Pkg)
-
-		if err != nil {
-			return fmt.Errorf("error in ingesting Pkg for PkgEqual: %v\n", err)
-		}
-
-		_, err = model.IngestPackage(ctx, client, *ingest.EqualPkg)
-
-		if err != nil {
-			return fmt.Errorf("error in ingesting EqualPkg for PkgEqual: %v\n", err)
-		}
-
-		_, err = model.PkgEqual(ctx, client, *ingest.Pkg, *ingest.EqualPkg, *ingest.PkgEqual)
-
-		if err != nil {
-			return fmt.Errorf("error in ingesting PkgEqual: %v\n", err)
-		}
-	}
-	return nil
-}
-
-func ingestHashEqual(ctx context.Context, client graphql.Client, graph assembler.IngestPredicates) error {
-	for _, ingest := range graph.HashEqual {
-		_, err := model.IngestArtifact(ctx, client, *ingest.Artifact)
-
-		if err != nil {
-			return fmt.Errorf("error in ingesting Artifact for HashEqual: %v\n", err)
-		}
-
-		_, err = model.IngestArtifact(ctx, client, *ingest.EqualArtifact)
-
-		if err != nil {
-			return fmt.Errorf("error in ingesting EqualArtifact for HashEqual: %v\n", err)
-		}
-
-		_, err = model.HashEqual(ctx, client, *ingest.Artifact, *ingest.EqualArtifact, *ingest.HashEqual)
-
-		if err != nil {
-			return fmt.Errorf("error in ingesting HashEqual: %v\n", err)
-		}
-	}
-	return nil
-}
-
-func ingestPointOfContact(ctx context.Context, client graphql.Client, graph assembler.IngestPredicates) error {
-	for _, ingest := range graph.PointOfContact {
-		var err error
-
-		if ingest.Src != nil {
-			_, err = model.IngestSource(ctx, client, *ingest.Src)
-		} else if ingest.Pkg != nil {
-			_, err = model.IngestPackage(ctx, client, *ingest.Pkg)
-
-		} else {
-			_, err = model.IngestArtifact(ctx, client, *ingest.Artifact)
-		}
-
-		if err != nil {
-			return fmt.Errorf("error in ingesting pkg/src/artifact PointOfContact: %v\n", err)
-		}
-
-		if ingest.Src != nil {
-			_, err = model.PointOfContactSrc(ctx, client, *ingest.Src, *ingest.PointOfContact)
-		} else if ingest.Pkg != nil {
-			_, err = model.PointOfContactPkg(ctx, client, *ingest.Pkg, ingest.PkgMatchFlag, *ingest.PointOfContact)
-		} else {
-			_, err = model.PointOfContactArtifact(ctx, client, *ingest.Artifact, *ingest.PointOfContact)
-		}
-
-		if err != nil {
-			return fmt.Errorf("error in ingesting PointOfContact: %v\n", err)
-		}
-	}
-	return nil
-}
-
-func ingestTestData(ctx context.Context, client graphql.Client, graph assembler.IngestPredicates) error {
-	if len(graph.IsDependency) > 0 {
-		err := ingestIsDependency(ctx, client, graph)
-		if err != nil {
-			return err
-		}
-	}
-
-	if len(graph.IsOccurrence) > 0 {
-		err := ingestIsOccurrence(ctx, client, graph)
-		if err != nil {
-			return err
-		}
-	}
-
-	if len(graph.HasSlsa) > 0 {
-		err := ingestHasSLSA(ctx, client, graph)
-		if err != nil {
-			return err
-		}
-	}
-
-	if len(graph.CertifyGood) > 0 {
-		err := ingestCertifyGood(ctx, client, graph)
-		if err != nil {
-			return err
-		}
-	}
-
-	if len(graph.HasSourceAt) > 0 {
-		err := ingestHasSourceAt(ctx, client, graph)
-		if err != nil {
-			return err
-		}
-	}
-
-	if len(graph.PkgEqual) > 0 {
-		err := ingestPkgEqual(ctx, client, graph)
-		if err != nil {
-			return err
-		}
-	}
-
-	if len(graph.HashEqual) > 0 {
-		err := ingestHashEqual(ctx, client, graph)
-		if err != nil {
-			return err
-		}
-	}
-
-	if len(graph.PointOfContact) > 0 {
-		err := ingestPointOfContact(ctx, client, graph)
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
 func Test_SearchSubgraphFromVuln(t *testing.T) {
 	server, err := startTestServer()
 
@@ -1351,7 +1101,7 @@ func Test_SearchSubgraphFromVuln(t *testing.T) {
 			expectedLen:       11,
 			expectedPkgs:      []string{"pkgType3", "pkgType2", "pkgType1", "pkgType4"},
 			expectedArtifacts: []string{"testArtifactAlgorithm1", "testArtifactAlgorithm2", "testArtifactAlgorithm3"},
-			graphInputs:       []assembler.IngestPredicates{simpleHasSLSAGraph, simpleIsDependencyGraph, isDependencyAndHasSLSARelationship},
+			graphInputs:       []assembler.IngestPredicates{simpleHasSLSAGraph, isDependencyAndHasSLSARelationship},
 		},
 		{
 			name:           "13: should not explore certifyGood case",
@@ -1457,7 +1207,7 @@ func Test_SearchSubgraphFromVuln(t *testing.T) {
 	for _, tt := range testCases {
 		t.Run(fmt.Sprintf("Test case %s\n", tt.name), func(t *testing.T) {
 			for _, graphInput := range tt.graphInputs {
-				err = ingestTestData(ctx, gqlClient, graphInput)
+				err = IngestTestData(ctx, gqlClient, graphInput)
 
 				if err != nil {
 					t.Errorf("error ingesting test data: %s", err)
@@ -1468,9 +1218,9 @@ func Test_SearchSubgraphFromVuln(t *testing.T) {
 			var getPackageIDsValues []*string
 			var startID string
 			if tt.startVersion != nil {
-				getPackageIDsValues, err = getPackageIDs(ctx, gqlClient, ptrfrom.String(tt.startType), tt.startNamespace, tt.startName, tt.startVersion, true, false)
+				getPackageIDsValues, err = GetPackageIDs(ctx, gqlClient, ptrfrom.String(tt.startType), tt.startNamespace, tt.startName, tt.startVersion, true, false)
 			} else {
-				getPackageIDsValues, err = getPackageIDs(ctx, gqlClient, ptrfrom.String(tt.startType), tt.startNamespace, tt.startName, nil, false, true)
+				getPackageIDsValues, err = GetPackageIDs(ctx, gqlClient, ptrfrom.String(tt.startType), tt.startNamespace, tt.startName, nil, false, true)
 			}
 
 			if err != nil {
@@ -1487,9 +1237,9 @@ func Test_SearchSubgraphFromVuln(t *testing.T) {
 			var stopID *string
 			if tt.stopType != nil {
 				if tt.stopVersion != nil {
-					getPackageIDsValues, err = getPackageIDs(ctx, gqlClient, tt.stopType, tt.stopNamespace, tt.stopName, tt.stopVersion, true, false)
+					getPackageIDsValues, err = GetPackageIDs(ctx, gqlClient, tt.stopType, tt.stopNamespace, tt.stopName, tt.stopVersion, true, false)
 				} else {
-					getPackageIDsValues, err = getPackageIDs(ctx, gqlClient, tt.stopType, tt.stopNamespace, tt.stopName, nil, false, true)
+					getPackageIDsValues, err = GetPackageIDs(ctx, gqlClient, tt.stopType, tt.stopNamespace, tt.stopName, nil, false, true)
 				}
 
 				if err != nil {
@@ -1505,7 +1255,7 @@ func Test_SearchSubgraphFromVuln(t *testing.T) {
 				stopID = getPackageIDsValues[0]
 			}
 
-			gotMap, err := SearchDependenciesFromStartNode(ctx, gqlClient, startID, stopID, tt.maxDepth)
+			gotMap, _, err := SearchDependenciesFromStartNode(ctx, gqlClient, startID, stopID, tt.maxDepth)
 
 			if err != nil {
 				t.Errorf("got err from SearchDependenciesFromStartNode: %s", err)
@@ -1517,7 +1267,7 @@ func Test_SearchSubgraphFromVuln(t *testing.T) {
 
 			var expectedPkgIDs []string
 			for _, pkg := range tt.expectedPkgs {
-				pkgIDs, err := getPackageIDs(ctx, gqlClient, &pkg, "", "", nil, false, false)
+				pkgIDs, err := GetPackageIDs(ctx, gqlClient, &pkg, "", "", nil, false, false)
 				if err != nil {
 					t.Errorf("expected package %s not found: %s\n", pkg, err)
 				}
@@ -1656,48 +1406,6 @@ func getGraphqlTestServer() (*handler.Server, error) {
 	srv := handler.NewDefaultServer(generated.NewExecutableSchema(config))
 
 	return srv, nil
-}
-
-// This function return matching packageName and/or packageVersion node IDs depending on if you specified to only find name nodes or version nodes
-func getPackageIDs(ctx context.Context, gqlClient graphql.Client, nodeType *string, nodeNamespace string, nodeName string, nodeVersion *string, justFindVersion bool, justFindName bool) ([]*string, error) {
-	var pkgFilter model.PkgSpec
-	if nodeVersion != nil {
-		pkgFilter = model.PkgSpec{
-			Type:      nodeType,
-			Namespace: &nodeNamespace,
-			Name:      &nodeName,
-			Version:   nodeVersion,
-		}
-	} else {
-		pkgFilter = model.PkgSpec{
-			Type: nodeType,
-		}
-	}
-
-	pkgResponse, err := model.Packages(ctx, gqlClient, pkgFilter)
-
-	if err != nil {
-		return nil, fmt.Errorf("error getting id for test case: %s\n", err)
-	}
-	var foundIDs []*string
-
-	if len(pkgResponse.Packages[0].Namespaces[0].Names) > 0 && !justFindVersion {
-		for _, name := range pkgResponse.Packages[0].Namespaces[0].Names {
-			foundIDs = append(foundIDs, &name.Id)
-		}
-	}
-
-	if len(pkgResponse.Packages[0].Namespaces[0].Names[0].Versions) > 0 && !justFindName {
-		for index := range pkgResponse.Packages[0].Namespaces[0].Names[0].Versions {
-			foundIDs = append(foundIDs, &pkgResponse.Packages[0].Namespaces[0].Names[0].Versions[index].Id)
-		}
-	}
-
-	if len(foundIDs) < 1 {
-		return nil, fmt.Errorf("no matching nodes found\n")
-	}
-
-	return foundIDs, nil
 }
 
 func getArtifactIDs(ctx context.Context, gqlClient graphql.Client, algorithm string) ([]string, error) {
